+++ tests/Unit/Caption/CollectionTypesCaptionTest.php
@@ -8,7 +8,7 @@ use Spatie\MediaLibrary\Models\Media;
 use Spatie\MediaLibrary\Tests\Support\TestModels\TestModel;
 use Spatie\MediaLibrary\Tests\TestCase;
 
-class CollectionTypesLegendTest extends TestCase
+class CollectionTypesCaptionTest extends TestCase
 {
     /**
      * @test
@@ -22,8 +22,8 @@ class CollectionTypesLegendTest extends TestCase
                 $this->addMediaConversion('thumb')->crop(Manipulations::CROP_CENTER, 60, 20);
             }
         };
-        $dimensionsLegendString = $testModel->mimeTypesLegend('logo');
-        $this->assertEquals('', $dimensionsLegendString);
+        $dimensionsCaptionString = $testModel->mimeTypesCaption('logo');
+        $this->assertEquals('', $dimensionsCaptionString);
     }
 
     /**
@@ -43,8 +43,8 @@ class CollectionTypesLegendTest extends TestCase
                 $this->addMediaConversion('thumb')->crop(Manipulations::CROP_CENTER, 60, 20);
             }
         };
-        $dimensionsLegendString = $testModel->mimeTypesLegend('logo');
-        $this->assertEquals('', $dimensionsLegendString);
+        $dimensionsCaptionString = $testModel->mimeTypesCaption('logo');
+        $this->assertEquals('', $dimensionsCaptionString);
     }
 
     /**
@@ -72,10 +72,10 @@ class CollectionTypesLegendTest extends TestCase
                 $this->addMediaConversion('thumb')->crop(Manipulations::CROP_CENTER, 100, 70);
             }
         };
-        $dimensionsLegendString = $testModel->mimeTypesLegend('logo');
-        $this->assertEquals(trans_choice('medialibrary::medialibrary.constraint.types', 3, [
+        $dimensionsCaptionString = $testModel->mimeTypesCaption('logo');
+        $this->assertEquals(trans_choice('medialibrary::medialibrary.constraints.types', 3, [
             'types' => 'jpeg, jpg, jpe, png',
-        ]), $dimensionsLegendString);
+        ]), $dimensionsCaptionString);
     }
 
     /**
@@ -103,9 +103,9 @@ class CollectionTypesLegendTest extends TestCase
                 $this->addMediaConversion('thumb')->crop(Manipulations::CROP_CENTER, 100, 70);
             }
         };
-        $dimensionsLegendString = $testModel->mimeTypesLegend('logo');
-        $this->assertEquals(trans_choice('medialibrary::medialibrary.constraint.types', 1, [
+        $dimensionsCaptionString = $testModel->mimeTypesCaption('logo');
+        $this->assertEquals(trans_choice('medialibrary::medialibrary.constraints.types', 1, [
             'types' => 'wav',
-        ]), $dimensionsLegendString);
+        ]), $dimensionsCaptionString);
     }
 }