+++ tests/Unit/Constraints/CollectionMimesValidationConstraintsTest.php
@@ -3,7 +3,6 @@
 namespace Spatie\MediaLibrary\Tests\Unit\Extension\UrlGenerator;
 
 use Spatie\Image\Manipulations;
-use Spatie\MediaLibrary\Exceptions\CollectionNotFound;
 use Spatie\MediaLibrary\File;
 use Spatie\MediaLibrary\Models\Media;
 use Spatie\MediaLibrary\Tests\Support\TestModels\TestModel;
@@ -30,7 +29,7 @@ class CollectionMimesValidationConstraintsTest extends TestCase
     /**
      * @test
      */
-    public function it_returns_mimes_validation_constraints_when_declared_in_collection()
+    public function it_returns_mimes_constraints_when_declared_in_collection()
     {
         $testModel = new class extends TestModel
         {
@@ -59,7 +58,7 @@ class CollectionMimesValidationConstraintsTest extends TestCase
     /**
      * @test
      */
-    public function it_returns_no_collection_mimes_validation_constraints_when_none_declared()
+    public function it_returns_no_collection_mimes_constraints_when_none_declared()
     {
         $testModel = new class extends TestModel
         {