+++ tests/Unit/Constraints/CollectionValidationConstraintsTest.php
@@ -12,7 +12,7 @@ class CollectionValidationConstraintsTest extends TestCase
     /**
      * @test
      */
-    public function it_returns_no_validation_constraint_when_non_existing_collection()
+    public function it_returns_no_constraint_when_non_existing_collection()
     {
         $testModel = new class extends TestModel
         {
@@ -21,52 +21,36 @@ class CollectionValidationConstraintsTest extends TestCase
                 $this->addMediaConversion('thumb')->crop(Manipulations::CROP_CENTER, 60, 20);
             }
         };
-        $validationConstraintsArray = $testModel->validationConstraints('logo');
-        $this->assertEquals([], $validationConstraintsArray);
+        $validationConstraints = $testModel->validationConstraints('logo');
+        $this->assertEquals([], $validationConstraints);
     }
 
     /**
      * @test
      */
-    public function it_returns_no_validation_constraint_with_non_existent_conversions()
+    public function it_returns_only_size_constraint_with_non_existent_conversions()
     {
-        $testModel = new class extends TestModel
-        {
-            public function registerMediaCollections()
-            {
-                $this->addMediaCollection('logo')->acceptsMimeTypes(['image/jpeg', 'image/png']);
-            }
-        };
-        $validationConstraintsArray = $testModel->validationConstraints('logo');
-        $this->assertEquals([], $validationConstraintsArray);
+        config()->set('medialibrary.max_file_size', 1000);
+        $validationConstraints = (new TestModel)->validationConstraints('logo');
+        $this->assertEquals(['max:1000'], $validationConstraints);
     }
 
     /**
      * @test
      */
-    public function it_returns_no_validation_constraint_when_none_is_declared()
+    public function it_can_return_no_constraint()
     {
-        $testModel = new class extends TestModel
-        {
-            public function registerMediaCollections()
-            {
-                $this->addMediaCollection('logo');
-            }
-
-            public function registerMediaConversions(Media $media = null)
-            {
-                $this->addMediaConversion('thumb');
-            }
-        };
-        $validationConstraintsArray = $testModel->validationConstraints('logo');
-        $this->assertEquals([], $validationConstraintsArray);
+        config()->set('medialibrary.max_file_size', null);
+        $validationConstraints = (new TestModel)->validationConstraints('avatar');
+        $this->assertEquals([], $validationConstraints);
     }
 
     /**
      * @test
      */
-    public function it_returns_only_dimension_validation_constraints_when_only_dimensions_declared()
+    public function it_can_return_only_dimension_constraints()
     {
+        config()->set('medialibrary.max_file_size', null);
         $testModel = new class extends TestModel
         {
             public function registerMediaCollections()
@@ -79,15 +63,16 @@ class CollectionValidationConstraintsTest extends TestCase
                 $this->addMediaConversion('thumb')->crop(Manipulations::CROP_CENTER, 60, 20);
             }
         };
-        $validationConstraintsArray = $testModel->validationConstraints('logo');
-        $this->assertEquals(['dimensions:min_width=60,min_height=20'], $validationConstraintsArray);
+        $validationConstraints = $testModel->validationConstraints('logo');
+        $this->assertEquals(['dimensions:min_width=60,min_height=20'], $validationConstraints);
     }
 
     /**
      * @test
      */
-    public function it_returns_only_mime_types_and_mimes_validation_constraints_when_only_mime_types_declared()
+    public function it_can_return_only_mime_types_and_mimes_constraints()
     {
+        config()->set('medialibrary.max_file_size', null);
         $testModel = new class extends TestModel
         {
             public function registerMediaCollections()
@@ -100,15 +85,26 @@ class CollectionValidationConstraintsTest extends TestCase
                 $this->addMediaConversion('thumb');
             }
         };
-        $validationConstraintsArray = $testModel->validationConstraints('logo');
-        $this->assertEquals(['mimes:jpeg,jpg,jpe,png', 'mimetypes:image/jpeg,image/png'], $validationConstraintsArray);
+        $validationConstraints = $testModel->validationConstraints('logo');
+        $this->assertEquals(['mimes:jpeg,jpg,jpe,png', 'mimetypes:image/jpeg,image/png'], $validationConstraints);
     }
 
     /**
      * @test
      */
-    public function it_returns_all_declared_validation_constraints()
+    public function it_can_return_only_size_constraint()
     {
+        config()->set('medialibrary.max_file_size', 1000);
+        $validationConstraints = (new TestModel)->validationConstraints('avatar');
+        $this->assertEquals(['max:1000'], $validationConstraints);
+    }
+
+    /**
+     * @test
+     */
+    public function it_can_return_all_constraints()
+    {
+        config()->set('medialibrary.max_file_size', 1000);
         $testModel = new class extends TestModel
         {
             public function registerMediaCollections()
@@ -121,10 +117,10 @@ class CollectionValidationConstraintsTest extends TestCase
                 $this->addMediaConversion('thumb')->crop(Manipulations::CROP_CENTER, 60, 20);
             }
         };
-        $validationConstraintsArray = $testModel->validationConstraints('logo');
+        $validationConstraints = $testModel->validationConstraints('logo');
         $this->assertEquals(
-            ['mimes:jpeg,jpg,jpe,png', 'mimetypes:image/jpeg,image/png', 'dimensions:min_width=60,min_height=20'],
-            $validationConstraintsArray
+            ['mimes:jpeg,jpg,jpe,png', 'mimetypes:image/jpeg,image/png', 'dimensions:min_width=60,min_height=20', 'max:1000'],
+            $validationConstraints
         );
     }
 }